% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check_output.R
\name{check_output}
\alias{check_output}
\title{Check output for warnings and errors}
\usage{
check_output(ff)
}
\arguments{
\item{ff}{path to a .html file output by `render()`}
}
\description{
Check output for warnings and errors
}

