% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_index.R
\name{check_rd_index}
\alias{check_rd_index}
\title{Check rd_index.yaml for missing or extra topics}
\usage{
check_rd_index(code_path = ".", rd_info = NULL, rd_index = NULL)
}
\arguments{
\item{code_path}{path to the source code directory of the package}

\item{rd_info}{used internally - do not specify if called interactively}

\item{rd_index}{used internally - do not specify if called interactively}
}
\description{
Check rd_index.yaml for missing or extra topics
}

