% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{deploy_travis}
\alias{deploy_travis}
\title{Deploy to Github Pages from Travis-CI}
\usage{
deploy_travis(email = "travis@travis-ci.org", name = "Travis CI")
}
\arguments{
\item{email}{email for commit}

\item{name}{name for commit}
}
\description{
Deploy to Github Pages from Travis-CI
}

