% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/function_ref_template.R
\name{function_ref_template}
\alias{function_ref_template}
\title{Generate the text to put in a rd.rmd file to build a package function reference}
\usage{
function_ref_template(package_name, code_path, exclude = NULL)
}
\arguments{
\item{package_name}{the name of the package, e.g. "packagedocs"}

\item{code_path}{path to the source code directory of the package}

\item{exclude}{vector of Rd entry names to exclude from the resulting document}
}
\description{
Generate the text to put in a rd.rmd file to build a package function reference
}

