% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_vignettes}
\alias{init_vignettes}
\title{Initialize a new packagedocs project}
\usage{
init_vignettes(code_path = ".", title = NULL, subtitle = NULL,
  author = NULL)
}
\arguments{
\item{code_path}{location of R package (defaults to current directory)}

\item{title}{title of main page (will search in DESCRIPTION if NULL - can be changed later)}

\item{subtitle}{subtitle of main page (can be changed later)}

\item{author}{author (can be changed later)}
}
\description{
Initialize a new packagedocs project
}

