% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{packagedocs_init}
\alias{packagedocs_init}
\alias{packagedocs_init_vignettes}
\title{Initialize a new packagedocs project}
\usage{
packagedocs_init(code_path = ".", docs_path = file.path(code_path, "docs"),
  package_name = NULL, title = NULL, subtitle = NULL, author = NULL,
  github_ref = NULL, index_file_rmd = "index.Rmd", rd_file_rmd = "rd.Rmd",
  build_file = TRUE)

packagedocs_init_vignettes(code_path = ".", title = NULL, subtitle = NULL,
  author = NULL)
}
\arguments{
\item{code_path}{location of R package (defaults to current directory)}

\item{docs_path}{location of code directory (defaults to a directory "docs" inside \code{code_path})}

\item{package_name}{the name of the package, e.g. "packagedocs" (will search in DESCRIPTION if NULL)}

\item{title}{title of main page (will search in DESCRIPTION if NULL - can be changed later)}

\item{subtitle}{subtitle of main page (can be changed later)}

\item{author}{author (can be changed later)}

\item{github_ref}{the "user/repo" part of a link to github - if NULL or "", the github link will not be displayed}

\item{index_file_rmd}{name of index file to be created}

\item{rd_file_rmd}{name of rd file to be created}

\item{build_file}{boolean to determine if the build file should be produced}
}
\description{
Initialize a new packagedocs project
}

