% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{packagedocs_init}
\alias{packagedocs_init}
\title{Initialize a new packagedocs project}
\usage{
packagedocs_init(code_path = ".", docs_path = file.path(code_path, "docs"),
  package_name = NULL, title = NULL, subtitle = "", author = NULL,
  github_ref = "user/repo")
}
\arguments{
\item{code_path}{location of docs directory}

\item{docs_path}{location of code directory (defaults to a directory "docs" inside \code{code_path})}

\item{package_name}{the name of the package, e.g. "packagedocs" (will search in DESCRIPTION if NULL)}

\item{title}{title of main page (will search in DESCRIPTION if NULL - can be changed later)}

\item{subtitle}{subtitle of main page (can be changed later)}

\item{author}{author (can be changed later)}

\item{github_ref}{the "user/repo" part of a link to github - if NULL or "", the github link will not be displayed}
}
\description{
Initialize a new packagedocs project
}

