% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purl_docs.R
\name{purl_docs}
\alias{purl_docs}
\title{Generate .R code files from docs}
\usage{
purl_docs(docs_base = "docs", code_base = "code")
}
\arguments{
\item{docs_base}{base diretory where all .Rmd files are located to scrape code from}

\item{code_base}{base directory of where to put resulting .R files}
}
\description{
Generate .R code files from docs
}

