% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rd_template.R
\name{rd_template}
\alias{rd_template}
\title{Generate the text to put in a rd.rmd file to build a package function reference}
\usage{
rd_template(package_name, code_path, rd_index = NULL, exclude = NULL)
}
\arguments{
\item{package_name}{the name of the package, e.g. "packagedocs"}

\item{code_path}{path to the source code directory of the package}

\item{rd_index}{path to yaml file with index layout information}

\item{exclude}{vector of Rd entry names to exclude from the resulting document}
}
\description{
Generate the text to put in a rd.rmd file to build a package function reference
}

