% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/render_docs.R
\name{render_docs}
\alias{render_docs}
\title{Render packagedocs}
\usage{
render_docs(code_path, docs_path, package_name, main_toc_collapse = TRUE,
  rd_toc_collapse = TRUE, lib_dir = "assets", render_main = TRUE,
  render_rd = TRUE, view_output = TRUE, rd_index = NULL)
}
\arguments{
\item{code_path}{location of docs directory}

\item{docs_path}{location of code directory}

\item{package_name}{the name of the package, e.g. "packagedocs"}

\item{main_toc_collapse}{use collapsing toc on main page}

\item{rd_toc_collapse}{use collapsing toc on rd page}

\item{lib_dir}{put assets in "assets" directory}

\item{render_main}{render main page}

\item{render_rd}{render rd page}

\item{view_output}{look at the output after render}

\item{rd_index}{optional path to rd layout yaml}
}
\description{
Render packagedocs
}

