% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_shell.R
\name{render_redirect}
\alias{render_redirect}
\title{Render vignette redirect}
\usage{
render_redirect(input_file_rmd, output_file_html)
}
\arguments{
\item{output_file_html}{output file name}

\item{intput_file_rmd}{input file name to read its yaml header with 'redirect' key}
}
\description{
Create a html redirect file that will automatically redirect to another webpage.  This is useful for making a html reidrect in a vignette.  By submitting a small redirect vignette, the package is not bloated in size
}

