% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_file.R
\name{test_fn_a}
\alias{test_fn_a}
\alias{test_fn_b}
\title{Test file}
\usage{
test_fn_a(a)

test_fn_b(b)
}
\arguments{
\item{a}{test param number 1}

\item{b}{test param number 2}
}
\description{
Test file
}
\examples{
test_fn_a(4)
test_fn_b("world")
hist(rnorm(1000))
}

