% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_docs.R
\name{vig_render_index}
\alias{index_render}
\alias{rd_render}
\alias{vig_render_index}
\alias{vig_render_rd}
\title{Render vignette files}
\usage{
vig_render_index(docs_path = "vignettes", code_path = ".",
  toc_collapse = TRUE, lib_dir = assets_dir(), render = TRUE,
  view_output = TRUE, is_https = FALSE, input_file_rmd = index_file_rmd(),
  output_file_html = index_file_html(),
  self_contained = is_self_contained_build(), cran_build = is_cran_build(),
  verbose = TRUE)

vig_render_rd(docs_path = "vignettes", code_path = ".",
  toc_collapse = FALSE, lib_dir = assets_dir(), render = TRUE,
  view_output = TRUE, rd_index = NULL, is_https = FALSE,
  input_file_rmd = rd_file_rmd(), temp_file_rmd = rd_temp_file_rmd(),
  output_file_html = rd_file_html(),
  self_contained = is_self_contained_build(), cran_build = is_cran_build(),
  verbose = TRUE)

rd_render(...)

index_render(...)
}
\arguments{
\item{docs_path}{location of index file}

\item{code_path}{location of library}

\item{toc_collapse}{boolean to determine if the table of contents should be collapsed}

\item{lib_dir}{directory where all assets are kept}

\item{render}{boolean to determine if the files should be rendered}

\item{view_output}{boolean to determine if a browser should be opened to the files}

\item{is_https}{boolean to determine if the cran http redirects should point to a http or https website}

\item{input_file_rmd}{rmd file input}

\item{output_file_html}{html file where the output is placed}

\item{self_contained}{boolean to determine if the html should be fully self contained}

\item{cran_build}{boolean to determine if the html should redirect to the github location of the package}

\item{verbose}{boolean to determine if ouput should be displayed}

\item{rd_index}{location of yaml file that describes the function references of the package.  Defaults to "rd_index.yaml"}

\item{temp_file_rmd}{temp rmd file that is a concatination of the input_file_rmd and the compiled rd_index.yaml file}

\item{...}{parameters passed directly to \code{vig_render_rd} or \code{vig_render_index}}
}
\description{
Render vignette files
}

