% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{build_vignettes}
\alias{build_vignettes}
\title{Build shell and gh-pages vignettes}
\usage{
build_vignettes(pkg = ".", dependencies = "VignetteBuilder",
  output_dir = "_gh-pages", devtools = FALSE)
}
\arguments{
\item{pkg}{path to package. Provided directly to \code{devtools::\link[as.package]()}}

\item{dependencies}{supplied directly to \code{devtools::\link[devtools]install_deps}()}

\item{output_dir}{directory where the fully contained vignette directory should be exported}

\item{devtools}{boolean to determine if the vignettes should be processed as self contained vignettes with devtools.  Runs \code{devtools::build_vignettes(pkg, dependencies)}}
}
\description{
Build shell html vignettes that are placed in inst/doc/ that redirect to the gh-pages branch of the github url provided in the DESCRIPTION file.  This function is heavily inspired by \code{devtools::\link[devtools]{build_vignettes}()}.
}

