% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travis.R
\name{deploy_travis}
\alias{deploy_travis}
\title{Deploy to Github Pages from Travis-CI}
\usage{
deploy_travis(repo = gsub(".*[:/]([^/]*/[^.]*)\\\\.git", "\\\\1",
  system("git config --get remote.origin.url", intern = TRUE)),
  valid_branches = "master", token_key = "GITHUB_TOKEN",
  email = "travis@travis-ci.org", name = "Travis CI", allow_pulls = FALSE,
  output_dir = "_gh-pages")
}
\arguments{
\item{repo}{character string that has the form USER/REPO}

\item{valid_branches}{branch name(s) that are allowed to deploy}

\item{token_key}{key name that will be autofilled}

\item{email}{email for commit}

\item{name}{name for commit}

\item{allow_pulls}{allow pull requests to process. (defaults to FALSE)}

\item{output_dir}{output directory to put the website in}
}
\description{
Deploy to Github Pages from Travis-CI
}

