% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\name{package_docs}
\alias{package_docs}
\alias{package_docs_rd}
\title{Generate package docs}
\usage{
package_docs(toc = TRUE, toc_depth = 2, toc_collapse = FALSE,
  extra_dependencies = NULL, self_contained = FALSE, fig_width = 6.5,
  fig_height = 4, mathjax = NULL, lib_dir = assets_dir(), ...,
  lazyrmd_render_fn = "html_document", lazyrmd_render_package = "rmarkdown")

package_docs_rd(..., rd_index = "rd_index.yaml", code_path = ".",
  exclude = NULL)
}
\arguments{
\item{toc}{should a table of contents be included?}

\item{toc_depth}{depth of the table of contents (max is 2 for this template)}

\item{toc_collapse}{should the table of contents have collapsible subsections?}

\item{extra_dependencies, self_contained, fig_width, fig_height, mathjax}{passed to the rmarkdown rendering function}

\item{lazyrmd_render_fn, lazyrmd_render_package, lib_dir}{arguments of \code{lazyrmd::\link[lazyrmd]{lazy_render}}.  Defaults to render with \code{rmarkdown::html_document}}

\item{rd_index, code_path, exclude}{parameters passed directly to \code{\link{rd_template}}}

\item{\ldots}{In \code{pacakge_docs}, the parameters are passed to the \code{lazyrmd::\link[lazyrmd]{lazy_render}} rendering function. In \code{package_docs_rd}, the parameters are passed to \code{package_docs}}
}
\description{
\code{rmarkdown} output formats for the main documentation and package reference pacakge respectively.
}

