% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_docs.R
\name{package_docs}
\alias{package_docs}
\title{Generate package docs}
\usage{
package_docs(toc = TRUE, toc_depth = 2, toc_collapse = FALSE,
  extra_dependencies = NULL, ...)
}
\arguments{
\item{toc}{should a table of contents be included?}

\item{toc_depth}{depth of the table of contents (max is 2 for this template)}

\item{toc_collapse}{should the table of contents have collapsible subsections?}

\item{extra_dependencies}{passed to \code{\link[rmarkdown]{html_document}}}

\item{\ldots}{parameters passed to \code{\link[rmarkdown]{html_document}}}
}
\description{
Generate package docs
}

