% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_template.R
\name{rd_template}
\alias{rd_template}
\title{Generate the text to put in a rd.rmd file to build a package function reference}
\usage{
rd_template(code_path, rd_index = NULL, exclude = NULL,
  run_examples = FALSE, img_path = rd_img_dir())
}
\arguments{
\item{code_path}{path to the source code directory of the package}

\item{rd_index}{path to yaml file with index layout information}

\item{exclude}{vector of Rd entry names to exclude from the resulting document}

\item{run_examples}{boolean to determine if examples should be run}
}
\description{
Generate the text to put in a rd.rmd file to build a package function reference
}

