% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_docs.R
\name{render_docs}
\alias{render_docs}
\title{Render packagedocs}
\usage{
render_docs(docs_path, code_path, package_name, main_toc_collapse = TRUE,
  rd_toc_collapse = FALSE, lib_dir = "assets", render_main = TRUE,
  render_rd = TRUE, view_output = TRUE, rd_index = NULL)
}
\arguments{
\item{docs_path}{location of R Markdown docs directory}

\item{code_path}{location of R package source directory}

\item{package_name}{the name of the package, e.g. "packagedocs"}

\item{main_toc_collapse}{use collapsing toc on main page}

\item{rd_toc_collapse}{use collapsing toc on rd page}

\item{lib_dir}{put assets in "assets" directory}

\item{render_main}{render main page}

\item{render_rd}{render rd page}

\item{view_output}{look at the output after render}

\item{rd_index}{optional path to rd layout yaml (if NULL, will search for "docs_path/rd_index.yaml" and use it if available)}
}
\description{
Render packagedocs
}

