% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rd.R
\name{render_rd}
\alias{render_rd}
\title{Generate rd.html file}
\usage{
render_rd(rd_skeleton, code_path, docs_path, rd_index = NULL,
  exclude = NULL, output_format = NULL, output = ".")
}
\arguments{
\item{rd_skeleton}{path to a .Rmd file that contains the header to use for the rd.Rmd}

\item{code_path}{path to the source code directory of the package}

\item{docs_path}{location of R Markdown docs directory}

\item{rd_index}{path to yaml file with index layout information}

\item{exclude}{vector of Rd entry names to exclude from the resulting document}

\item{output_format}{passed to \code{\link[rmarkdown]{render}} - \code{\link{package_docs}} is used by default}

\item{output}{directory to put the output rd.Rmd and rd.html file}
}
\description{
Generate rd.html file
}

